requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0 